hook file for createDefaultLearningPath, HandleSubmit added
